% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_cendat.R
\docType{data}
\name{can_cendat}
\alias{can_cendat}
\title{Population Data from 2011/2016 Canadian Census Tracts}
\format{
A data.frame

\describe{
\itemize{
\item Dimensions: 303 rows and 26 columns
\item Data attributes:
\itemize{
\item CDUID: Uniquely identifies a census division (composed of the 2-digit province or territory unique identifier followed by the 2-digit census division code)
\item CDNAME: Census division name
\item CDTPE: Census division type
\item PRNAME: Province Name
\item PRUID: Province Unique Identifier
\item pop.16: Total Population in 2016
\item pop.11:  Total Population in 2011
\item pop.chg: Percent Change in Population 2011-2016
\item dwel.16: Total Private Dwellings 2016
\item dwel.11: Total Private Dwellings 2016
\item dwel.chg: Percent Change in Number of Private Dwellings 2011-2016
\item landarea.sqkm: Land Area in Square Kilometres 2016
\item popdens.sqkm: Population density pe square kilometre 2016
\item poprank.nat: National population rank 2016
\item poprank.prov: Provincial/territorial population rank 2016
}
}
}
}
\source{
https://open.canada.ca/data/en/dataset/ece81c43-aa4e-41ef-86c2-3835eb5aa95c
}
\usage{
can_cendat
}
\description{
Population and dwelling counts, for Canada and
census divisions, 2016 and 2011 censuses
}
\keyword{datasets}
