% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_shp.R
\docType{data}
\name{can_shp}
\alias{can_shp}
\title{Shapefile of Canada}
\format{
An sf simple feature collection with 1 multipolygon

\describe{
\itemize{
\item Dimensions: 1 features and 0 fields
\item Projection: Un-projected
\item Coordinate Reference System: NAD83
\item Geographic Extent: Canada
}
}
}
\source{
https://open.canada.ca/data/en/dataset/5be03a46-8504-40a7-a96c-af195bae0428
}
\usage{
can_shp
}
\description{
Dissolved Canadian Provinces with 200-meter simplification
}
\keyword{datasets}
