% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easblu_on.R
\docType{data}
\name{easblu_on}
\alias{easblu_on}
\title{Annual Frequency of Eastern Bluebird in Ontario}
\format{
A RasterLayer

\describe{
\itemize{
\item Dimensions: 761, 1010, 768610  (nrow, ncol, ncell)
\item Resolution: 2-4 kilometers
\item Projection: Albers equal-area projection
\item Geographic Extent: Ontario
\item Data: Full-Year Occurence Probability (the expected probability (0-1) of occurrence a species)
}
}
}
\source{
https://ebird.github.io/ebirdst/
}
\usage{
easblu_on
}
\description{
Data from the eBird Status and Trends Dataset (Version 2021),
by the Cornell Lab of Ornithology (Fink et al., 2022).
Obtained with the ebirdst package
}
\keyword{datasets}
